Seperated Trait Mono-morphization into a separate file, moved some functions around. Significantly improved documentation.
