+++ src/core/mod.rs
@@ -1,3 +1,7 @@
+//! Contains various components and structures supporting the creation of a chessboard. This
+//! includes `SQ`, `BitBoard`, `Player`, `Piece`, `GenTypes`, `Rank`, and `File`. Also holds
+//! the statically created `MagicHelper`, which at runtime creates various lookup tables.
+
 #[macro_use]
 mod bit_macros;
 
@@ -5,16 +9,47 @@ pub mod bit_twiddles;
 pub mod piece_move;
 pub mod magic_helper;
 pub mod masks;
-
+pub mod mono_traits;
 pub mod sq;
 pub mod bitboard;
-
-use std::fmt;
 
 use self::bit_twiddles::*;
 use self::masks::*;
 use self::sq::SQ;
-use self::bitboard::BitBoard;
+
+use std::fmt;
+
+
+/// Array of all possible pieces, indexed by their enum value.
+pub const ALL_PIECES: [Piece; PIECE_CNT] =
+    [Piece::P, Piece::N, Piece::B, Piece::R, Piece::Q, Piece::K];
+
+/// Array of both players, indexed by their enum value.
+pub const ALL_PLAYERS: [Player; 2] = [Player::White, Player::Black];
+
+/// Array of all `Files`s, indexed by their enum value.
+pub const ALL_FILES: [File; FILE_CNT] = [
+    File::A,
+    File::B,
+    File::C,
+    File::D,
+    File::E,
+    File::F,
+    File::G,
+    File::H,
+];
+
+/// Array of all `Rank`s, indexed by their enum value.
+pub const ALL_RANKS: [Rank; RANK_CNT] = [
+    Rank::R1,
+    Rank::R2,
+    Rank::R3,
+    Rank::R4,
+    Rank::R5,
+    Rank::R6,
+    Rank::R7,
+    Rank::R8,
+];
 
 
 /// Enum to represent the Players White & Black.
@@ -41,6 +76,7 @@ impl Player {
         sq ^ SQ((*self) as u8 * 56)
     }
 
+    /// Gets the direction of a pawn push for a given player.
     #[inline]
     pub fn pawn_push(&self) -> i8 {
         match *self {
@@ -49,21 +85,45 @@ impl Player {
         }
     }
 
+    /// Returns the relative rank of a square in relation to a player.
+    ///
+    ///  # Examples
+    ///
+    /// ```rust
+    /// use pleco::core::{Player,Rank};
+    /// use pleco::core::sq::SQ;
+    ///
+    /// let w = Player::White;
+    /// let b = Player::Black;
+    ///
+    /// assert_eq!(w.relative_rank_of_sq(SQ::A1), Rank::R1);
+    /// assert_eq!(b.relative_rank_of_sq(SQ::H8), Rank::R1);
+    /// assert_eq!(b.relative_rank_of_sq(SQ::A1), Rank::R8);
+    /// ```
     #[inline]
     pub fn relative_rank_of_sq(&self, sq: SQ) -> Rank {
         self.relative_rank(sq.rank_of_sq())
     }
 
+    /// Returns the relative rank of a rank in relation to a player.
+    ///
+    ///  # Examples
+    ///
+    /// ```rust
+    /// use pleco::core::{Player,Rank};
+    /// use pleco::core::sq::SQ;
+    ///
+    /// let w = Player::White;
+    /// let b = Player::Black;
+    ///
+    /// assert_eq!(w.relative_rank(Rank::R1), Rank::R1);
+    /// assert_eq!(b.relative_rank(Rank::R8), Rank::R1);
+    /// assert_eq!(b.relative_rank(Rank::R1), Rank::R8);
+    /// ```
     #[inline]
     pub fn relative_rank(&self, rank: Rank) -> Rank {
         ALL_RANKS[((rank as u8) ^ (*self as u8 * 7)) as usize]
     }
-}
-
-
-#[inline]
-pub fn relative_rank(p: Player, rank: Rank) -> Rank {
-    ALL_RANKS[((rank as u8) ^ (p as u8 * 7)) as usize]
 }
 
 
@@ -81,131 +141,20 @@ impl fmt::Display for Player {
     }
 }
 
-
-pub trait PlayerTrait {
-    fn player() -> Player;
-    fn opp_player() -> Player;
-
-    fn down(sq: SQ) -> SQ;
-    fn up(sq: SQ) -> SQ;
-    fn left(sq: SQ) -> SQ;
-    fn right(sq: SQ) -> SQ;
-
-    fn down_left(sq: SQ) -> SQ;
-    fn down_right(sq: SQ) -> SQ;
-    fn up_left(sq: SQ) -> SQ;
-    fn up_right(sq: SQ) -> SQ;
-
-    fn shift_down(bb: BitBoard) -> BitBoard;
-    fn shift_up(bb: BitBoard) -> BitBoard;
-    fn shift_left(bb: BitBoard) -> BitBoard;
-    fn shift_right(bb: BitBoard) -> BitBoard;
-
-    fn shift_down_left(bb: BitBoard) -> BitBoard;
-    fn shift_down_right(bb: BitBoard) -> BitBoard;
-    fn shift_up_left(bb: BitBoard) -> BitBoard;
-    fn shift_up_right(bb: BitBoard) -> BitBoard;
-}
-
-pub struct WhiteType {}
-pub struct BlackType {}
-
-impl PlayerTrait for WhiteType {
-    fn player() -> Player {
-        Player::White
-    }
-    fn opp_player() -> Player {
-        Player::Black
-    }
-
-    fn down(sq: SQ) -> SQ { sq - SQ(8) }
-
-    fn up(sq: SQ) -> SQ { sq + SQ(8) }
-
-    fn left(sq: SQ) -> SQ { sq - SQ(1) }
-
-    fn right(sq: SQ) -> SQ { sq + SQ(1) }
-
-    fn down_left(sq: SQ) -> SQ { sq - SQ(9) }
-
-    fn down_right(sq: SQ) -> SQ { sq - SQ(7) }
-
-    fn up_left(sq: SQ) -> SQ { sq + SQ(7) }
-
-    fn up_right(sq: SQ) -> SQ { sq + SQ(9) }
-
-    fn shift_down(bb: BitBoard) -> BitBoard { bb >> 8 }
-
-    fn shift_up(bb: BitBoard) -> BitBoard { bb << 8 }
-
-    fn shift_left(bb: BitBoard) -> BitBoard { (bb & !BitBoard::FILE_A) >> 1 }
-
-    fn shift_right(bb: BitBoard) -> BitBoard { (bb & !BitBoard::FILE_H) << 1 }
-
-    fn shift_down_left(bb: BitBoard) -> BitBoard { (bb & !BitBoard::FILE_A) >> 9 }
-
-    fn shift_down_right(bb: BitBoard) -> BitBoard { (bb & !BitBoard::FILE_H) >> 7 }
-
-    fn shift_up_left(bb: BitBoard) -> BitBoard { (bb & !BitBoard::FILE_A) << 7 }
-
-    fn shift_up_right(bb: BitBoard) -> BitBoard { (bb & !BitBoard::FILE_H) << 9 }
-}
-
-impl PlayerTrait for BlackType {
-    fn player() -> Player {
-        Player::Black
-    }
-
-    fn opp_player() -> Player {
-        Player::White
-    }
-
-    fn down(sq: SQ) -> SQ { sq + SQ(8) }
-
-    fn up(sq: SQ) -> SQ { sq - SQ(8) }
-
-    fn left(sq: SQ) -> SQ { sq + SQ(1) }
-
-    fn right(sq: SQ) -> SQ { sq - SQ(1) }
-
-    fn down_left(sq: SQ) -> SQ { sq + SQ(9) }
-
-    fn down_right(sq: SQ) -> SQ { sq + SQ(7) }
-
-    fn up_left(sq: SQ) -> SQ { sq - SQ(7) }
-
-    fn up_right(sq: SQ) -> SQ { sq - SQ(9) }
-
-    fn shift_down(bb: BitBoard) -> BitBoard { bb << (8) }
-
-    fn shift_up(bb: BitBoard) -> BitBoard { bb >> (8) }
-
-    fn shift_left(bb: BitBoard) -> BitBoard {
-        (bb & !BitBoard::FILE_H) << (1)
-    }
-
-    fn shift_right(bb: BitBoard) -> BitBoard {
-        (bb & !BitBoard::FILE_A) >> (1)
-    }
-
-    fn shift_down_left(bb: BitBoard) -> BitBoard {
-        (bb & !BitBoard::FILE_H) << (9)
-    }
-
-    fn shift_down_right(bb: BitBoard) -> BitBoard {
-        (bb & !BitBoard::FILE_A) << (7)
-    }
-
-    fn shift_up_left(bb: BitBoard) -> BitBoard {
-        (bb & !BitBoard::FILE_H) >> (7)
-    }
-
-    fn shift_up_right(bb: BitBoard) -> BitBoard {
-        (bb & !BitBoard::FILE_A) >> (9)
-    }
-}
-
-/// Publicly available move-generation types.
+/// Different types of move generating options.
+///
+///`GenTypes::All` -> All available moves.
+///`GenTypes::Captures` -> All captures and both capture/non-capture promotions.
+///`GenTypes::Quiets` -> All non captures and both capture/non-capture promotions.
+///`GenTypes::QuietChecks` -> Moves likely to give check.
+///`GenTypes::Evasions` -> Generates evasions for a board in check.
+///`GenTypes::NonEvasions` -> Generates all moves for a board not in check.
+///
+///  # Safety
+///
+/// `GenTypes::QuietChecks` and `GenTypes::NonEvasions` can only be used if the board
+/// if not in check, while `GenTypes::Evasions` can only be used if the the board is
+/// in check. The remaining `GenTypes` can be used legally whenever.
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub enum GenTypes {
     All,
@@ -216,52 +165,7 @@ pub enum GenTypes {
     NonEvasions
 }
 
-pub trait GenTypeTrait {
-    fn gen_type() -> GenTypes;
-}
 
-pub struct AllGenType {}
-pub struct CapturesGenType {}
-pub struct QuietsGenType {}
-pub struct QuietChecksGenType {}
-pub struct EvasionsGenType {}
-pub struct NonEvasionsGenType {}
-
-impl GenTypeTrait for AllGenType {
-    fn gen_type() -> GenTypes {
-        GenTypes::All
-    }
-}
-
-impl GenTypeTrait for CapturesGenType {
-    fn gen_type() -> GenTypes {
-        GenTypes::Captures
-    }
-}
-
-impl GenTypeTrait for QuietsGenType {
-    fn gen_type() -> GenTypes {
-        GenTypes::Quiets
-    }
-}
-
-impl GenTypeTrait for QuietChecksGenType {
-    fn gen_type() -> GenTypes {
-        GenTypes::QuietChecks
-    }
-}
-
-impl GenTypeTrait for EvasionsGenType {
-    fn gen_type() -> GenTypes {
-        GenTypes::Evasions
-    }
-}
-
-impl GenTypeTrait for NonEvasionsGenType {
-    fn gen_type() -> GenTypes {
-        GenTypes::NonEvasions
-    }
-}
 
 /// Enum for all the possible Pieces.
 #[repr(u8)]
@@ -292,8 +196,6 @@ impl Piece {
 
 }
 
-
-
 impl fmt::Display for Piece {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         let s = match *self {
@@ -307,61 +209,6 @@ impl fmt::Display for Piece {
         f.pad(s)
     }
 }
-
-pub const ALL_PIECES: [Piece; PIECE_CNT] =
-    [Piece::P, Piece::N, Piece::B, Piece::R, Piece::Q, Piece::K];
-
-pub trait PieceTrait {
-    fn piece_type() -> Piece;
-}
-
-pub struct PawnType {}
-pub struct KnightType {}
-pub struct BishopType {}
-pub struct RookType {}
-pub struct QueenType {}
-pub struct KingType {}
-
-impl PieceTrait for PawnType {
-    fn piece_type() -> Piece {
-        Piece::P
-    }
-}
-
-impl PieceTrait for KnightType {
-    fn piece_type() -> Piece {
-        Piece::N
-    }
-}
-
-impl PieceTrait for BishopType {
-    fn piece_type() -> Piece {
-        Piece::B
-    }
-}
-
-impl PieceTrait for RookType {
-    fn piece_type() -> Piece {
-        Piece::R
-    }
-}
-
-impl PieceTrait for QueenType {
-    fn piece_type() -> Piece {
-        Piece::Q
-    }
-}
-
-impl PieceTrait for KingType {
-    fn piece_type() -> Piece {
-        Piece::K
-    }
-}
-
-
-pub const ALL_PLAYERS: [Player; 2] = [Player::White, Player::Black];
-
-
 
 /// Enum for the Files of a Chessboard.
 #[repr(u8)]
@@ -391,80 +238,77 @@ pub enum Rank { // eg a specific row
     R8 = 7,
 }
 
-
-pub const ALL_FILES: [File; FILE_CNT] = [
-    File::A,
-    File::B,
-    File::C,
-    File::D,
-    File::E,
-    File::F,
-    File::G,
-    File::H,
-];
-
-pub const ALL_RANKS: [Rank; RANK_CNT] = [
-    Rank::R1,
-    Rank::R2,
-    Rank::R3,
-    Rank::R4,
-    Rank::R5,
-    Rank::R6,
-    Rank::R7,
-    Rank::R8,
-];
-
-
-/// Types of Castling available
+/// Types of Castling available to a player.
 #[derive(Copy, Clone, PartialEq, Debug)]
 pub enum CastleType {
     KingSide = 0,
     QueenSide = 1,
 }
 
-
-// For whatever rank the bit is in, gets the whole bitboard
+/// For whatever rank the bit (inner value of a `SQ`) is, returns the
+/// corresponding rank as a u64.
 #[inline]
 pub fn rank_bb(s: u8) -> u64 {
     RANK_BB[rank_of_sq(s) as usize]
 }
 
+/// For whatever rank the bit (inner value of a `SQ`) is, returns the
+/// corresponding `Rank`.
 #[inline]
 pub fn rank_of_sq(s: u8) -> Rank {
     ALL_RANKS[(s >> 3) as usize]
 }
 
+/// For whatever rank the bit (inner value of a `SQ`) is, returns the
+/// corresponding `Rank` index.
 #[inline]
 pub fn rank_idx_of_sq(s: u8) -> u8 {
     (s >> 3) as u8
 }
 
+/// For whatever file the bit (inner value of a `SQ`) is, returns the
+/// corresponding file as a u64.
 #[inline]
 pub fn file_bb(s: u8) -> u64 {
     FILE_BB[file_of_sq(s) as usize]
 }
 
+/// For whatever file the bit (inner value of a `SQ`) is, returns the
+/// corresponding `File`.
 #[inline]
 pub fn file_of_sq(s: u8) -> File {
     ALL_FILES[(s & 0b0000_0111) as usize]
 }
 
+/// For whatever file the bit (inner value of a `SQ`) is, returns the
+/// corresponding `File` index.
 #[inline]
 pub fn file_idx_of_sq(s: u8) -> u8 {
     (s & 0b0000_0111) as u8
 }
 
-// Assumes only one bit!
+/// Converts a singular bit of a u64 to it's index in the u64.
+/// If there's more than one bit in the u64, this will be done for
+/// the least significant bit.
+///
+/// # Safety
+///
+/// Undefined behavior if there are 0 bits in the input.
 #[inline]
 pub fn u64_to_u8(b: u64) -> u8 {
     debug_assert_eq!(popcount64(b), 1);
     bit_scan_forward(b)
 }
 
-// Given a Square (u8) that is valid, returns the bitboard representaton
+/// Given a square (u8) that is valid, returns the bitboard representation
+/// of that square.
+///
+/// # Safety
+///
+/// If the input is greater than 63, an empty u64 will be returned.
 #[inline]
 pub fn u8_to_u64(s: u8) -> u64 {
-    assert!(s < 64);
+    debug_assert!(s < 64);
     (1 as u64).wrapping_shl(s as u32)
 }
